+++ air/src/execution_step/trace_handler/handler.rs
@@ -160,11 +160,9 @@ impl TraceHandler {
     }
 
     pub(crate) fn fold_end_with_error(&mut self, fold_id: &str) {
-        // unwrap here is used because this function must be called from a fold block with
-        // corresponding fold_id and since it's a error handling, it's better not to produce
-        // a new error
         let fold_fsm = match self.fsm_keeper.extract_fold(fold_id) {
             Ok(fold_fsm) => fold_fsm,
+            // just passing here is ok, because error could be produced while fold initialization
             Err(_) => return,
         };
         fold_fsm.fold_end_with_error(&mut self.data_keeper);
